fix: type errors
First we type-assert the result of npmFetch.json to match its usage. We can assume that it is either an array or record of packages.
We also move the log until later in the code and log the objects variable instead of results, since that is guarantied to be an array. We also convert the array to a string using join for better logging