misc: change bin structure
Instead of building the package using tsc and then running the js file, we can run using ts-node. That way we only need ts files in the project. For this to work a few changes needed to be done:
- Removed the initial empty index file
- Moved bin-files into lib
- Converted bin files to ts and renamed them to index/index-cn
- Changed the shebang in the index files to use ts-node. This also requires ts-node is a dependency of the package
- Adjust tsconfig in order to work with ts-node

Seems like the package still works as intended. It might take a little bit longer to run because of the additional compile step