refactor: extract utility function
There are many points in the code where a package name is combined with a version using `${name}@${version}` or similar. Extracted this logic to a utility function. This process also showed that a potentially undefined version was used in some prints. Changed logic to only print version where it is defined.