Split core module (#58)
* refactor: make property required

Make version property required in dependency. This makes sense as Unity dependencies always give their dependency versions explicitly. The only reason this was set to optional was because it was not immediately initialized at one point.

* refactor: extract utility function

There are many points in the code where a package name is combined with a version using `${name}@${version}` or similar. Extracted this logic to a utility function. This process also showed that a potentially undefined version was used in some prints. Changed logic to only print version where it is defined.

* refactor: move and refactor function

Extract function for getting latest version of pkg-info to own module, so it can be grouped with other pkg-info related functions. Also change name to tryGetLatestVersion to indicate that this function may return undefined. Also refactor the function a bit for readability

a

* chore: reformat file

Just add a line break for consistency

* chore: reformat files

* refactor: move function

Move isInternalPackage function into package-name module to be closer to other related functions

* fix: incorrect type

This function assumes the name includes no version

* refactor: move function

Move manifest related functions from core to own module

* chore: clean up imports and reformat

* refactor: move functions

Move upm-config related functions to own module

* refactor: extract module

Move env related code into own module

* refactor: rename type

This type does not refer to a generic semantic version but specifically to a unity-editor version

* refactor: extract module

Extract editor-version related logic to own module

* refactor: move function

Move function to get search options into cmd-search since that is the only place where it is used

* refactor: move functions

Move functions for interacting with npm registry into client module.

* refactor: rename module