chore: better ts-config
Previously we had a separate ts-config for tests. This was because we don't want test code to be emitted to lib.

This however caused issues with WebStorm, at least for me, because the IDE was not capable of correctly tracking two ts-configs live.

The fix is to use one ts-config for the whole projects (including tests), but then have a separate ts-config just for building, which extends the first one, but excludes tests.

Inspired by: https://stackoverflow.com/a/60223923