refactor: null instead of error (#90)
* refactor: remove error from function

Refactor editor-version comparison logic to not throw an error anymore but instead return null. Also add a version of the function which accepts pre-parsed versions.

Adjust tests to work with this new syntax, as well as adding new ones to test for when null should be returned.

* refactor: remove error from function

Make `decodeBasicAuth` return null if decoding was unsuccessful instead of throwing an error. Also renamed to `tryDecodeBasicAuth` to reflect this change.

Adjus tests and add a new one to test correct null behaviour.