refactor: rename package / manifest types (#95)
* refactor: extract package-manifest type

Add type which represents the content of a unity package manifest (package.json)

* refactor: adopt npm naming terms

Rename pkg-info to packument, as this is the term used by npm for the equivalent. Put Unity before, so there is not ambiguity.

PkgInfo -> UnityPackument
PkgVersionInfo -> UnityPackumentVersion

* refactor: project manifest type

Type was called pkg-manifest, which could be misleading because the package manifest is usually the `package.json` while this type refers to `manifest.json`. Type was renamed to project-manifest.

Also added a few properties which were missing from the documentation.