refactor: extract npm client param (#100)
* refactor: make comment jsdoc

* doc: document function params

* refactor: extract parameter

Make `fetchPackument` function take the npm-client as a parameter. That way it will not be recreated for each call.

* refactor: make comment jsdoc

* doc: document function params

* refactor: make function param required

The function gains nothing by the param being undefiend. Make it required.

* refactor: extract parameter

Make `fetchPackageDependencies` function take the npm-client as a parameter. That way it will not be recreated for each call.