refactor: npm-client file-structure and types (#161)
* refactor: rename file

Rename file to match the name of primary type exported from it.

* refactor: rename function

the `getNpmClient` function is effectively the constructor function for the npm-client. Renamed it to `makeNpmClient` to better indicate this.

* refactor: split file

The logic for resolving dependencies was in the npm-client file. The file was getting pretty large because of this and also mixed levels of abstraction.

Moved dependency-resolving logic to own file

* refactor: make types readonly

Make npm-client related types immutable