refactor: no logs in manifest-io (#168)
* refactor: pull up manifest-not-found log

The error for when a manifest is not found is now no longer logged inside project-manifest-io, but inside the cmd functions that use it. This is one step in a series of refactors to separate domain logic and cli display logic.

* refactor: pull up manifest-parse-error log

The error for when a manifest could not be parsed is now no longer logged inside project-manifest-io, but inside the cmd functions that use it. This is one step in a series of refactors to separate domain logic and cli display logic.

* refactor: pull up manifest-save-error log

The error for when a manifest could not be saved is now no longer logged inside project-manifest-io, but inside the cmd functions that use it. This is one step in a series of refactors to separate domain logic and cli display logic.

* refactor: remove unused imports

* refactor: simplify tests

We can remove mock-console from project-manifest-io tests since we no longer test for logs there.

* refactor: extract function

Extract a function for logging manifest-load errors.

* refactor: extract function

Extract a function for logging manifest-save errors.