refactor: extract project-version load function (#169)
* refactor: extract utility function

Extract function for determining path of ProjectVersion.txt based on project directory path.

* refactor: make error more generic

The ManifestParseError was specifically for when manifests could not be parsed. Made this error more generic for when a file could not be parsed in general.

* refactor: make property optional

Some parse errors don't have a cause.

* refactor: extract utility function

Extract function for loading project editor-version.