refactor: remove check-path
The `parseEnv` function had the option to check some of the file-system dependent checks. This was only used by tests. After removing this logic it seemed that only the specific tests which concerned themselves with the check-path logic failed.

Remove the check-path logic and relevant tests. To make sure this does not break anything a following commit will make the env tests more rigorous.