refactor: merge properties
Currently Env had the `upstreamRegistry` property plus the `upstream` property on it. `upstreamRegistry` was always filled and then `upstream` indicated whether it should be used.

This information can be encoded more directly by making `upstreamRegistry` optional. Now if it is null, it is clear we should not use it. This also makes it easier for a future change which will allow to use n registries.