test: domain-name prop-tests  (#201)
* feat: domain-name generator

* test: improve package-id validation test

Use prop-tests for tests that are not dependent on a concrete input value.

* test: improve package-reference validation tests

Use prop-tests for tests which don't depend on a concrete input value.

* test: add more test examples

* test: improve package-url validation tests

Use prop tests for validation.

* test: better examples

* test: improve manifest tests

Make dependency add/remove tests property-based.

* test: improve manifest tests

Use property-based tests in testables tests

* test: improve scoped-registry tests

Use property-based tests with packument-name where applicable.

* refactor: rename file

In `fast-check`, the library we are using for property-based tests, data-generators are called arbitraries. In order to match this convention, renamed the module which generates domain-names to `domain-name.arb.ts`.