refactor: extract npm-token update service (#215)
* refactor: use explicit return type

* refactor: move function

Move npm-token update function to npmrc-io module.

* refactor: rename function

* refactor: pull up log

Move log up by one function

* refactor: add error type

* feat: handle error

The result returned by `tryUpateAuthToken` was not used. Previously this function simply threw on error, so that is why. Now the result is checked and bubbled-up if it is an error.

* refactor: auth-token update

Convert to service function. The function uses io instead of being io itself. it should thus not be in the io module.

* refactor: rename function

* test: add tests