refactor: change folder structure
Currently, CLI-related code is somewhat spread throughout the project. In my opinion it is better to concentrate this logic into a single place.

This change moves all logic that is strictly cli related into a cli directory. Also renames the previous `cli.ts` file to `index.ts` since it now represents the directory.