refactor: get built-in packages (#225)
* refactor: rename module

We had a module which contained the logic for determining the users home path. Now I want to add another function for finding a special path. In order to not have many modules for all these paths, all with just one function, we rename the `home` module to `special-paths` which can contain all of these.

* feat: get path to editor-install

Adds the `tryGetEditorInstall` function which can be used to get the default installation path to a specific editor install.

* feat: get directory names in directory

* feat: get builtin packages