refactor: strong child-process error (#230)
* refactor: add error types

Add custom error types or more accurately type-aliases for error types.

* refactor: use better error type

* refactor: shorten

* refactor: explicit error type

* refactor: strong child-process error

`ChildProcessError`was just an alias to `childProcess.ExecException` which intern was basically just `Error` with a few extra optional properties. This caused `ChildProcessError` to basically be equal to `Error` which prevented TS  from doing some type-checking magic as it was equal to every other error type in the project (because they are also all subtypes of `Error`).

To prevent this made `ChildProcessError` a proper class.