refactor: adjust typing
The jest-matcher for errors does not carry any generic information about the result that was matched on. Thus we really cannot know if the "Error" type of the result is actually a js `Error`. For example, in the case of the `add` function it is actually a `Error[]`.

To better reflect this, make the type of the parameter a generic. This way clients of this function can fill in themselves what the type is, if required.