refactor: cmd-add manifest io (#233)
* refactor: cmd-add manifest io

Currently, when calling `add` with multiple packages, each will load, modify and save the project-manifest. This repeated loading and saving was necessary in case a package caused the app to crash, so that previous modifications would still be saved.

For a while now, the app should no longer crash in regular use. Instead it will return error objects. This means we can now safely load and save the manifest only once.

* fix: missing log