feat: atomic package add (#234)
Currently, when adding multiple packages using the `add` command, each will be added separately. That means, that if packages A, B and C are added and B fails for any reason, A and C will still be added to the manifest.

With this change add operations are now atomic. In the same scenario as before, the manifest would not be modified, since one of the packages could not be added.

A side-effect of this change is that `add` now only returns one error, instead of an array of them, since the first encountered error cancels the operation.

See #223