refactor: no fs-based tests (#246)
* refactor: manifest save spy

When spying on `trySaveProjectManifest` it should not call the actual implementation.

* test: decouple from file-system

Remove usages of logic that uses the real file-system for tests.

* refactor: remove module usage

The only place where the `unity-project` module was still used was in `env.test`, where it's root-path constant was used.

Since this is the only usage, simply create a local constant.

* refactor: drop unused logic

File-system based unity-project mocking is no longer used.

* refactor: drop unused logic

`process.env` mocking was no longer used. Removed it.