refactor: improve string parsing (#279)
* refactor: improve string-parsing

String parsing functions now propagate the error that caused the failure, allowing for better logging.

Also extracted common logic into a utility function.

* refactor: rename file

The data-parsing module is actually just concerned with parsing strings. Renamed to better match this.

* test: rename test

The term `notify` is mostly used in this project to indicate logging. Since there is no logging in this test, the name does not really fit.

* test: add toml parsing tests