fix: incorrect log
The error log for when a package version could not be found is currently incorrect.

When adding package `com.some.package@2.0.0` and the primary registry has the package, just not that version the app will still fail with "not found". This is because `com.some.package` does not exist at all in the upstream registry and that error overrides the first "version not found".

Fixed by checking which error  is more fixable and printing that. A missing version is more fixable than the package not existing at all.