refactor: make function generic (#304)
* refactor: make function more generic

The `tryGetTargetEditorVersionFor` only acts on properties which are part of `UnityPackageManifest`. It does not require all the properties which come with `UnityPackumentVersion`. We can thus make the function more generic by allowing it to work on all package-manifests.

* refactor: move function

Since the `tryGetTargetEditorVersionFor` function now acts on package-manifests it makes sense to move it into the corresponding file.

* tests: make tests generic

In accordance with the previous changes, make the tests use package-manifests instead of packument-versions

* refactor: make function more generic

Make the `tryGetTargetEditorVersionFor` function more generic by only requiring the input parameter to have the properties it actually uses for the function.