refactor: merge dependency props
Currently resolved dependencies have two properties indicating their source:

- `internal` indicates whether the package is built-in
- `source` indicates the source url.

These properties should not be orthogonal. For example, a builtin-package simply does not have a source-url, so for packages with `internal: true` the `source` property should be missing.

To remedy these issues merge the two properties. Now the `source` property can have a special "built-in" value to indicate a internal/built-in package.