refactor: make function more focused
The `tryResolveFromPackument` function should be responsible for resolving a specific version from a packument. But currently it's signature also includes errors it can't really produce as well as having context, like the source, injected into it only to produce it as output.

This is not really necessary. We should keep this function simple and add context in the client functions. Refactored for this purpose.