fix: bad declaration file
The declaration files for `npm-registry-fetch` package do not expose the error-class it throws when a fetch failed. Because of this, we wrote our own declaration file for it.

The declaration file was incorrectly written. It placed the error class into the packages `index` file even though it is inside `lib/errors`. After updating the declaration and relevant imports the error is now fixed.

See #319.