refactor: resolve packument service (#335)
This PR adds a service for resolving packuments by searching a list of source registries in order. For now this is only used in cmd-view (which has the nice effect of making the cmd function cli-logic only), but it could also be used in other places in the future, especially when it comes to #49.