refactor: upmconfig load logic
Currently we have a service function for determining the directory in which the upmconfig is located. On load and save the actual file path of the config is constructed from this directory.

There is no need for this delayed path construction. We can instead make the service resolve the actual path of the config file right away. This makes client code a little shorter.