refactor: clean up upm auth
Currently we return the path of the upm config file we modified after a save so that the path can be in logging. This is done because previously we only had access to the configs directory path in the client code.

Since #337 we also have access to the file path in client code and so dont need this return value.