refactor: separate user and debug logs (#342)
Separates the logging functions for user-facing and debug logs. User-facing logs should only be made on the CLI layer while debug logs can be made everywhere. This way low-level errors which only get logged to the debug log don't need to be bubbled all the way up to the CLI layer.