refactor: extract service (#346)
Currently we always parse the projects editor version so that we can put it into the env object. But we only use the editor-version inside the add command. So there is no reason to load it for example for the search command.

This change extracts the logic for determining a projects editor-version into its own service which is now used by the add command. This way users can now use other commands outside of unity projects.