refactor: child-process service function (#355)
* refactor: extract service function

Make the logic to run child processes a service-style function. This way it is more in line with the rest of the app.

* refactor: move file

The run-child-process function acts like an io function. Move it to the corresponding folder

* refactor: rename module

Rename the module for running child processes to "child-process" to better reflect this.