refactor: dependency resolving logic
Currently, you can either pass a semantic version or "latest" to the dependency-resolve service. The service will then resolve the packages latest version if no semantic version was given.

This complicates the logic inside the service. For this reason, now the client calling the service needs to determine the latest version and pass it to the service.