refacor: split upm config type (#379)
Currently the UPMConfig type represents the file content of a upmconfig.toml file. Since the datatype has the same structure as the file, the whole application has to be somewhat designed around it, which can be cumbersome.

This change splits this type into two, one (UpmConfigContent) still represents the file, but UpmConfig now only holds the auth information in a easier to access format. This makes it`s usage simpler in the service layer of the app. The conversion now happens in the load/put services which interact with the upm config.