refactor: remove unused properties (#380)
When designing some of the types which represent external data, we added all properties which exist in that data even when we don't use them.  This makes interacting and mocking these data types more difficult than it needs to be.

This change removes these properties since they have no impact on the app at all.