refactor: simplify cmd object type
The cmd object type stores global options inside the `_global` property. This is a leftover from the javascript days of this project. With typescript help we dont need to nest the global properties like this and can pull them up into the options object. Intellisense will help to discover these properties