refactor: singleton services (#381)
Currently we create all services inside the cli/index.ts file. This clutters the file and also makes it hard to get any specific service in a place where it can't be injected.

Since all our service functions are singletons we can instantiate them in place, with default dependencies injected and thus get them more easily. But since the build functions for the service function are still there, we can still mock dependencies in tests.