misc: reword log (#384)
The log in question was printed when the upmconfig does not contain an entry for the primary registry. This is usually expected, since usually the primary registry is openupm and you dont need auth for it.
Still you would get a warning that openupm could not be parsed. But actually this log just means that there was no entry.

The log was now reworded and also is only shown on the verbose level. Also we now only show this log when using a custom registry, since, again, openupm will never need auth.