refactor: upmconfig registry logic (#385)
Currently we get the auth information for registries on startup, regardless of whether they will be used. This has a small performance penalty, but mostly it introduces some software architecture issues, by coupling all commands to the parse-env logic and upmconfig.

This change makes it so that the config is only loaded on demand and cached. We also don't parse the whole config, only requested entries.