refactor: simplify error type
Currently we use the `HttpErrorBase` from the `npm-registry-client` library for http errors. We needed to create the type definitions for this type ourself since the original is JS only. But really we only ever need the statusCode property off this type. So it is far easier to create an interface type with exactly that requirement.

This change adds that type and adjusts the code everywhere accordingly