refactor: extract utility function
When interacting with npm registries we handle the case where the response code is 401 specially because we throw a different error in that case. We could consider this business logic that is duplicated in these io functions.

This change extracts that common logic into a separate function that can be reused and tested.