refactor: inject io only (#391)
Currently we inject all sorts of functions into all sorts of functions. After reading a few articles about dependency injection, I have come to the conclusion that it would be better if we only inject IO function and build all other services in-place or just inline them.

We now have four layers in the app (which I will also document in a dev guide soon)

- Domain: Pure logic related to the problem domain
- IO: Functions for interacting with external systems, like the network or the fs.
- App: Utility functions which combine IO functions into reusable compositions.
- CLI: CLI specific code.

Sources for this decision were the great articles by Mark Seemann:
- https://blog.ploeh.dk/2017/02/02/dependency-rejection/
- https://blog.ploeh.dk/2017/01/30/partial-application-is-dependency-injection/
- https://blog.ploeh.dk/2017/01/27/dependency-injection-is-passing-an-argument/
- https://blog.ploeh.dk/2017/01/27/from-dependency-injection-to-dependency-rejection/
- https://blog.ploeh.dk/2017/01/03/decoupling-application-errors-from-domain-models/
- https://blog.ploeh.dk/2023/10/02/dependency-whac-a-mole/