deps: drop another-npm-registry-client
`another-npm-registry-client` is a fork of `npm-registry-client` which is a, now deprecated, library for interacting with npm registries. At this point, the project does not use any runtime code from the library (see 893a67eb7108ac23ef715bfa898bf61ab884f52c and a678e14d804ba6ec8fc2df22f250f6d9e82092ae). The only part of it we still used was the NpmAuth type which represents npm authentication data. But there is already a type for this as part of the `npm-registry-fetch` libary. Therefore the dependency can now safely be removed.