fix: Allow any number for minor version number (#423)
Currently only numbers 1-4 are allowed in the minor version number.
Since Unity 6, also the number 0 is used (6000.0). This commit
changes the regex for the minor version number to include any number,
which fixes support for Unity 6 and is more permissive in case any
other numbers are used by Unity in the future.